Refactor code and delete unneeded dependencies
